Add destruction of files
